// src/app/models/eventos/PutEventosModel.ts
export class PutEventosModel{
    idEvento: string = '';
    dia_hora: string = '';
    nome_evento: string = '';
    descricao: string = '';
    carga_horaria: string = '';   
    organizadores: string = '';    
    palestrante: string = '';  
    valor: string = '';
    status: string = '';
    local: string = '';
    }