// src/app/models/eventos/GetEventosModel.ts
export class GetEventosModel{
    idEvento: string = '';
    dia_hora: string = '';
    nome_evento: string = '';
    descricao: string = '';
    carga_horaria: string = '';   
    organizadores: string = '';    
    palestrante: string = '';  
    valor: string = '';
    status: string = '';
    local: string = '';
    }